move call functions to query
