+++ query/src/functions/mod.rs
@@ -13,5 +13,7 @@
 // limitations under the License.
 
 mod context_function;
+pub mod systems;
 
 pub use context_function::ContextFunction;
+pub use systems::FunctionFactory;
